refector: pages 폴더 구조 간단 변경
