--- src/pages/call/CallPage.tsx
@@ -1,5 +1,0 @@
-import { VideoCallWidget } from '@/widgets/video-call/VideoCallWidget';
-
-export default function CallPage() {
-  return <VideoCallWidget />;
-}
