--- src/pages/chat/AiChatPage.tsx
@@ -1,5 +1,0 @@
-import { AiChatWidget } from '@/widgets/chat/AiChatWidget';
-
-export default function AiChatPage() {
-  return <AiChatWidget />;
-}
