--- src/pages/chat/ChatPage.tsx
@@ -1,5 +1,0 @@
-import { ChatWidget } from '@/widgets/chat/ChatWidget';
-
-export default function ChatPage() {
-  return <ChatWidget />;
-}
