--- src/pages/users/UsersPage.tsx
@@ -1,5 +1,0 @@
-import { UserListWidget } from '@/widgets/user/UserListWidget';
-
-export default function UsersPage() {
-  return <UserListWidget />;
-}
